Pass error objects into callbacks
First step: conforming to node.js standards. Errors, if not surfaced
as separate events to bind to, should be passed as the first parameter
to allow the callback to deal with it meaningfully.

Currently, only the errors from the HTTP layer directly are being passed
(timeouts and such). Further checks can be added to allow callees to
distinguish between success and failure without data.status == 'success'
checks (which, notably, fails in coupon creation).