Refactor requests to better capture errors
Errors from HTTP, or failed statuses, will now be sent back to the callback. Additionally, the request-response handling code is now more centralized in the process_response() function, which optionally can add per-request error checking, but checks for the "typical" errors without any supervision.

So now, if the host is down, or the request fails, this can be messages appropriately to the user in a consistent fashion.